+++ lttng/org.eclipse.linuxtools.tmf.core.tests/src/org/eclipse/linuxtools/tmf/core/tests/trace/indexer/AbstractCheckpointCollectionTest.java
@@ -257,9 +257,9 @@ public abstract class AbstractCheckpointCollectionTest {
         for (int i = 0; i < CHECKPOINTS_INSERT_NUM; i++) {
             if (random) {
                 Random rand = new Random();
-                list.add(rand.nextInt(CHECKPOINTS_INSERT_NUM));
+                list.add(Integer.valueOf(rand.nextInt(CHECKPOINTS_INSERT_NUM)));
             } else {
-                list.add(i);
+                list.add(Integer.valueOf(i));
             }
         }
         return list;
@@ -279,7 +279,8 @@ public abstract class AbstractCheckpointCollectionTest {
 
         for (int i = 0; i < CHECKPOINTS_INSERT_NUM; i++) {
             Integer randomCheckpoint = list.get(i);
-            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(12345 + randomCheckpoint), new TmfLongLocation(123456L + randomCheckpoint), 0);
+            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(12345 + randomCheckpoint.intValue()),
+                    new TmfLongLocation(123456L + randomCheckpoint.intValue()), 0);
             long found = fCheckpointCollection.binarySearch(checkpoint);
             assertEquals(randomCheckpoint.intValue(), found);
         }
@@ -306,9 +307,9 @@ public abstract class AbstractCheckpointCollectionTest {
         for (int i = 0; i < CHECKPOINTS_INSERT_NUM; i++) {
             if (random) {
                 Random rand = new Random();
-                list.add(rand.nextInt(CHECKPOINTS_INSERT_NUM));
+                list.add(Integer.valueOf(rand.nextInt(CHECKPOINTS_INSERT_NUM)));
             } else {
-                list.add(i);
+                list.add(Integer.valueOf(i));
             }
         }
 
@@ -318,7 +319,8 @@ public abstract class AbstractCheckpointCollectionTest {
 
         for (int i = 0; i < CHECKPOINTS_INSERT_NUM; i++) {
             Integer randomCheckpoint = list.get(i);
-            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(12345), new TmfLongLocation(123456L + randomCheckpoint), 0);
+            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(12345),
+                    new TmfLongLocation(123456L + randomCheckpoint.intValue()), 0);
             long found = fCheckpointCollection.binarySearch(checkpoint);
             assertEquals(randomCheckpoint.intValue(), found);
         }
@@ -353,12 +355,14 @@ public abstract class AbstractCheckpointCollectionTest {
     public void testBinarySearchInBetweenSameTimestamp() {
         int checkpointNum = 0;
         for (; checkpointNum < 100; checkpointNum++) {
-            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(0), new TmfLongLocation((long) checkpointNum), checkpointNum);
+            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(0),
+                    new TmfLongLocation(checkpointNum), checkpointNum);
             fCheckpointCollection.insert(checkpoint);
         }
 
         for (; checkpointNum < 200; checkpointNum++) {
-            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(1), new TmfLongLocation((long) checkpointNum), checkpointNum);
+            TmfCheckpoint checkpoint = new TmfCheckpoint(new TmfTimestamp(1),
+                    new TmfLongLocation(checkpointNum), checkpointNum);
             fCheckpointCollection.insert(checkpoint);
         }
