+++ lttng/org.eclipse.linuxtools.tmf.core/src/org/eclipse/linuxtools/tmf/core/ctfadaptor/CtfTmfEvent.java
@@ -75,7 +75,7 @@ public class CtfTmfEvent extends TmfEvent
 
         fDeclaration = declaration;
         sourceCPU = cpu;
-        typeId = declaration.getId();
+        typeId = declaration.getId().longValue();
         eventName = declaration.getName();
 
     }
@@ -189,7 +189,7 @@ public class CtfTmfEvent extends TmfEvent
         if (getContent() != null) {
             ITmfEventField ipField = getContent().getField(CtfConstants.CONTEXT_FIELD_PREFIX + CtfConstants.IP_KEY);
             if (ipField != null && ipField.getValue() instanceof Long) {
-                long ip = (Long) ipField.getValue();
+                long ip = ((Long) ipField.getValue()).longValue();
                 callsite = getTrace().getCTFTrace().getCallsite(eventName, ip);
             }
         }
